django/res/requirements.txt
Django>=4
django-cors-headers>=4.3.1
requests>=2.32.3
psycopg2>=2.9.9
djangorestframework>=3.12
djangorestframework-simplejwt>=4.8.0
PyJWT>=2.1,<3
channels>=4.0.0
daphne>=4.0.0
asgiref>=3.6.0
